[RCM] Add remote config support to the tracing library
Adds the ability for subsystems running within the tracer to utilize
remote configuration. If a product is configured in the tracer that
utilizes remote configuration, a remote configuration client is created
and started. The client polls periodically in the background, handlings
updates from the agent and dispatching configuration updates to the
relevant subsystems.